Refactor NDArray to use a two-parameter design NDArray<T, M extends Marker>
- Introduce Marker interface and concrete marker classes: Float32Marker, Float64Marker, Complex64Marker, Complex128Marker, Int32Marker, Int64Marker, BooleanMarker.
- Update NDArray class definition to `final class NDArray<T, M extends Marker>`.
- Update DType to be generic on T and M: `DType<T, M extends Marker>`.
- Refactor all operations under lib/src/operations/ to use the new 2-parameter signatures.
- Extract operators extension into lib/src/operators.dart to avoid circular imports.
- Update all tests under test/ to use the new 2-parameter type system, including cleaning up obsolete wrapper constructor calls.
- Fix generic factory type inference issues by using explicit type arguments where necessary to avoid context-type fallback to Marker.
- All 834 tests are passing successfully.

TAG=agy
CONV=e7d8ce00-bb8d-4815-91d9-45e218b58da3

+++ pkgs/ndarray/test/shape_manipulation_test.dart
@@ -25,7 +25,7 @@ void main() {
           expect(b.toList(), [1.0, 2.0, 3.0, 4.0]);
 
           // Verify it shares memory (view behavior)
-          b.data[0] = Float64(99.0);
+          b.data[0] = 99.0;
           expect(a.data[0], 99.0);
         }),
       );
@@ -113,7 +113,7 @@ void main() {
           expect(b.toList(), [1.0, 2.0]);
 
           // View verification
-          b.data[0] = Float64(42.0);
+          b.data[0] = 42.0;
           expect(a.data[0], 42.0);
         }),
       );
@@ -219,7 +219,7 @@ void main() {
           expect(b.toList(), [1.0, 4.0, 2.0, 5.0, 3.0, 6.0]);
 
           // View verification
-          b.data[0] = Float64(11.0); // updates a.data[0]
+          b.data[0] = 11.0; // updates a.data[0]
           expect(a.data[0], 11.0);
         }),
       );
@@ -315,7 +315,7 @@ void main() {
           expect(b.toList(), [1.0, 2.0, 1.0, 2.0, 1.0, 2.0]);
 
           // Verify it allocates new memory (copy behavior)
-          b.data[0] = Float64(99.0);
+          b.data[0] = 99.0;
           expect(a.data[0], 1.0);
         }),
       );
@@ -391,7 +391,7 @@ void main() {
           expect(b.toList(), [1.0, 1.0, 2.0, 2.0, 3.0, 3.0]);
 
           // Copy verification
-          b.data[0] = Float64(99.0);
+          b.data[0] = 99.0;
           expect(a.data[0], 1.0);
         }),
       );
@@ -494,7 +494,7 @@ void main() {
           expect(b.dtype, DType.float64);
           expect(b.toList(), [10.0, 20.0, 30.0]);
 
-          b.data[0] = Float64(99.0);
+          b.data[0] = 99.0;
           expect(a.data[0], 10.0);
         }),
       );
@@ -516,7 +516,7 @@ void main() {
           expect(b.isContiguous, true);
           expect(b.toList(), [1.0, 3.0, 2.0, 4.0]);
 
-          b.data[0] = Float64(99.0);
+          b.data[0] = 99.0;
           expect(parent.data[0], 1.0);
         }),
       );
@@ -540,7 +540,7 @@ void main() {
           expect(b.dtype, DType.float64);
           expect(b.toList(), [10.0, 20.0, 30.0]);
 
-          b.data[0] = Float64(99.0);
+          b.data[0] = 99.0;
           expect(a.data[0], 10.0);
         }),
       );
